Set linecolor attribute for function used to display cursor
Max 7 reduces the number of color attributes available for function
objects. Previously, the function object used to display the moving
cursor had all colors set to transparent with pointcolor set to white.
In Max 7, linecolor alone replaces linecolor, pointcolor, and 
sustaincolor, so for the cursor to be visible, linecolor must also be
set to white. (Because the cursor is a single point the line itself is
never drawn in any case.)