Create 264.line breakpoint function module (#17)
Closes #6 

* Create 264.line breakpoint function module

* Add 264.line to README

* Add 264.line to unreleased changes in CHANGELOG

* Start line playback from scrub position and follow scrub during playback

Previously, playback always started from x=0, now it starts from the
last set scrub position (and resets scrub position to x=0 when it is
stopped/done playing). In addition, scrubbing during playback now moves
the playhead. This allows a user to scrub to x=n during playback and
playback will continue from n.

* Remove default @loop 0 in 264.line

`patcherargs @loop 0` would trigger after a loadbang in a parent patch
causing a module sent “loop 1” on load to still have looping disabled

* Add more detail to 264.line help file

* Set linecolor attribute for function used to display cursor

Max 7 reduces the number of color attributes available for function
objects. Previously, the function object used to display the moving
cursor had all colors set to transparent with pointcolor set to white.
In Max 7, linecolor alone replaces linecolor, pointcolor, and 
sustaincolor, so for the cursor to be visible, linecolor must also be
set to white. (Because the cursor is a single point the line itself is
never drawn in any case.)

* Add 2px space between copy and paste buttons in function editor

* Add auto-sizing to 264.line

* Allow programmatic generation of random lines

* Add “reverse” method to 264.line

* Allow lines to be cleared programmatically

* Add undo/redo functionality to 264.line editor

* Fix line input for 264.line

* Update messages tab of 264.line help patcher to include new features

* Add 264.line to the Module Finder extra

* Scale randomly generated lines to full x-axis range (0–127)

* Add “move mode” to 264.line function editor

* Show editor line in yellow when in “move mode”

The line in the function editor (and its points) will be coloured yellow 
to match the colour of the “move mode” button in its active state to 
help indicate the editor’s state

* Display hand/grab cursor when over function editor in “move mode”

To help users know they are in move mode, this shows a custom “hand” 
cursor when over the function editor. While the mosue is pressed, the 
hand closes as if it is grabbing the function.

* Add shortcut to close function editor with escape key

* Refactor 264.line to make function editor more maintainable

* Add a draw/pencil mode to the 264.line editor

* Document 264.line messages in reference file

* Update parameter catalogue

* Add 264.line to init & interfaces

* Add tab to 264.line help file explaining use with presets

* Tweak 264.line GUI layout and function editor window style

* Update README screenshot to include 264.line

* Improve hint that clicking on small function in 264.line opens editor